requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
blinker==1.7.0
Brotli==1.1.0
cached-property==1.5.2
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cryptography==41.0.5
cycler==0.12.1
decorator==5.1.1
drjit==0.4.3
executing==2.0.1
flatbuffers==23.5.26
fonttools==4.44.0
frozenlist==1.4.0
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.3
h5py==3.10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipydatawidgets==4.3.2
ipython==8.12.3
ipywidgets==8.0.5
jax==0.4.2
jedi==0.19.1
joblib==1.3.2
jupyterlab-widgets==3.0.5
keras==2.13.1
kiwisolver==1.4.5
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mitsuba==3.4.0
multidict==6.0.4
numpy==1.24.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pip==23.3.1
prompt-toolkit==3.0.39
protobuf==4.21.12
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pythreejs==2.4.2
pyu2f==0.1.5
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
setuptools==68.2.2
sionna==0.15.1
six==1.16.0
stack-data==0.6.3
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorflow==2.13.1
tensorflow-estimator==2.13.0
termcolor==2.3.0
threadpoolctl==3.2.0
traitlets==5.13.0
traittypes==0.2.1
typing_extensions==4.5.0
urllib3==2.0.7
wcwidth==0.2.9
Werkzeug==3.0.1
wheel==0.41.3
widgetsnbextension==4.0.9
wrapt==1.15.0
yarl==1.9.2
zipp==3.17.0
